more cleaning, fix a bunch of asserts, remove unused soevideocapture crap, upgrade libs to latest: libxml2, libjpeg, perforce, pcre, zlib

+++ src/engine/client/application/Direct3d9/src/win32/Direct3d9_DynamicVertexBufferData.cpp
@@ -257,9 +257,9 @@ int Direct3d9_DynamicVertexBufferData::getNumberOfLockableDynamicVertices(bool w
 
 // ----------------------------------------------------------------------
 
-int Direct3d9_DynamicVertexBufferData::getSortKey()
+intptr_t Direct3d9_DynamicVertexBufferData::getSortKey()
 {
-	return reinterpret_cast<int>(ms_d3dVertexBuffer);
+	return reinterpret_cast<intptr_t>(ms_d3dVertexBuffer);
 }
 
 // ----------------------------------------------------------------------
