+++ src/external/3rd/library/platform/utils/Base/win32/Thread.h
@@ -81,7 +81,7 @@ namespace Base
 	    protected:
 		    bool            mThreadContinue;
         private:
-		    uint32  		mThreadID;
+		    uintptr_t  		mThreadID;
             bool            mThreadActive;
         };
